requirements.txt
dateparser
mistune
rq>=1.12.0
django~=4.2.4
django-auditlog
django-markdownx @ git+https://github.com/alphatownsman/django-markdownx.git@e69480c64ad9c5d0499f4a8625da78cf2bb7691b
django-jsonform
libsass
django-compressor
django-sass-processor
django-rq
django-simple-history  # no longer used but kept for migration code which is used by tests
django-auditlog @ git+https://github.com/jazzband/django-auditlog.git@45591463e8192b4ac0095e259cc4dcea0ac2fd6c
django-hijack
django-user-messages
django-slack
django-ninja
django-polymorphic
django-maintenance-mode
django-tz-detect
django-bleach
django-redis
django-oauth-toolkit
django-anymail
easy-thumbnails
lxml
openpyxl
psycopg2-binary
requests
filetype
setproctitle
tqdm
opencc
dnspython
typesense
markdownify
igdb-api-v4
podcastparser
listparser
fontawesomefree
discord.py
loguru
